Rename `poses_accessor.py` to `move_accessor.py`
